Revert "Battery pulling from backend"
This reverts commit c7e7d9972573df4a6b7e0859abfc9f7c861f0e21.

+++ SUITS_MagicLeap_Implemented_/Assets/VitalScripts/SuitVitalBar.cs
@@ -8,14 +8,14 @@ public class SuitVitalBar : MonoBehaviour
     public Gradient gradient;
     public Image vitalStatusBar;
 
-    public void setMaxVitalsValue(float vitalStatus)
+    public void setMaxVitalsValue(int vitalStatus)
     {
        slider.maxValue = vitalStatus;
        slider.value = vitalStatus;
        vitalStatusBar.color =  gradient.Evaluate(1f);
     }
 
-    public void setVitalsValue(float vitalStatus)
+    public void setVitalsValue(int vitalStatus)
     {
         slider.value = vitalStatus;
         vitalStatusBar.color = gradient.Evaluate(slider.normalizedValue);
